requirements.txt
flask
flask-sqlalchemy
flask-login
flask-migrate
flask-wtf
markdown
bleach
python-dotenv